regen classic payments in one class
