+++ lib/adyen/services/payment.rb
@@ -1,23 +1,128 @@
-require_relative 'payment/general_api'
-require_relative 'payment/modifications_api'
-
+require_relative './service'
 module Adyen
-  class Payment
+  class Payment < Service
     attr_accessor :service, :version
 
     DEFAULT_VERSION = 68
     def initialize(client, version = DEFAULT_VERSION)
-      @service = 'Payment'
-      @client = client
-      @version = version
+      super(client, version, 'Payment')
     end
 
-    def general_api
-      @general_api ||= Adyen::GeneralApi.new(@client, @version)
+    def adjust_authorisation(request, headers: {})
+      endpoint = '/adjustAuthorisation'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
     end
 
-    def modifications_api
-      @modifications_api ||= Adyen::ModificationsApi.new(@client, @version)
+    def authorise(request, headers: {})
+      endpoint = '/authorise'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def authorise3d(request, headers: {})
+      endpoint = '/authorise3d'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def authorise3ds2(request, headers: {})
+      endpoint = '/authorise3ds2'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def cancel(request, headers: {})
+      endpoint = '/cancel'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def cancel_or_refund(request, headers: {})
+      endpoint = '/cancelOrRefund'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def capture(request, headers: {})
+      endpoint = '/capture'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def donate(request, headers: {})
+      endpoint = '/donate'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def get_authentication_result(request, headers: {})
+      endpoint = '/getAuthenticationResult'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def refund(request, headers: {})
+      endpoint = '/refund'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def retrieve3ds2_result(request, headers: {})
+      endpoint = '/retrieve3ds2Result'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def technical_cancel(request, headers: {})
+      endpoint = '/technicalCancel'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
+    end
+
+    def void_pending_refund(request, headers: {})
+      endpoint = '/voidPendingRefund'.gsub(/{.+?}/, '%s')
+      endpoint = endpoint.gsub(%r{^/}, '')
+      endpoint = format(endpoint)
+      
+      action = { method: 'post', url: endpoint }
+      @client.call_adyen_api(@service, action, request, headers, @version)
     end
 
   end
