--- lib/adyen/services/payment/general_api.rb
@@ -1,56 +1,0 @@
-require_relative '../service'
-module Adyen
-  class GeneralApi < Service
-    attr_accessor :service, :version
-
-    def initialize(client, version = DEFAULT_VERSION)
-      super(client, version, 'Payment')
-    end
-
-    def authorise(request, headers: {})
-      endpoint = '/authorise'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def authorise3d(request, headers: {})
-      endpoint = '/authorise3d'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def authorise3ds2(request, headers: {})
-      endpoint = '/authorise3ds2'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def get_authentication_result(request, headers: {})
-      endpoint = '/getAuthenticationResult'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def retrieve3ds2_result(request, headers: {})
-      endpoint = '/retrieve3ds2Result'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-  end
-end
