--- lib/adyen/services/payment/modifications_api.rb
@@ -1,83 +1,0 @@
-require_relative '../service'
-module Adyen
-  class ModificationsApi < Service
-    attr_accessor :service, :version
-
-    def initialize(client, version = DEFAULT_VERSION)
-      super(client, version, 'Payment')
-    end
-
-    def adjust_authorisation(request, headers: {})
-      endpoint = '/adjustAuthorisation'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def cancel(request, headers: {})
-      endpoint = '/cancel'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def cancel_or_refund(request, headers: {})
-      endpoint = '/cancelOrRefund'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def capture(request, headers: {})
-      endpoint = '/capture'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def donate(request, headers: {})
-      endpoint = '/donate'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def refund(request, headers: {})
-      endpoint = '/refund'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def technical_cancel(request, headers: {})
-      endpoint = '/technicalCancel'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-    def void_pending_refund(request, headers: {})
-      endpoint = '/voidPendingRefund'.gsub(/{.+?}/, '%s')
-      endpoint = endpoint.gsub(%r{^/}, '')
-      endpoint = format(endpoint)
-      
-      action = { method: 'post', url: endpoint }
-      @client.call_adyen_api(@service, action, request, headers, @version)
-    end
-
-  end
-end
